Extract Telegram Utilities
